Remove some temporary vars in Tests
Xcode 6 beta 2 now allows the removal of some of the temporary
variables which were required to perform comparisons, but not all.

This is a bit confusing as many of these have the same return type.
I'm not sure why comparing a function which returns an array to an
array works in some cases but not others, it made more sense when
it didn't work for any of them.